+++ opower-rest-client-builder/src/test/java/com/opower/rest/test/Codahale3MetricsIntegrationTesting.java
@@ -2,6 +2,7 @@ package com.opower.rest.test;
 
 import com.codahale.metrics.MetricRegistry;
 import com.codahale.metrics.SharedMetricRegistries;
+import com.google.common.base.Throwables;
 import com.netflix.config.ConfigurationManager;
 import com.netflix.hystrix.HystrixCommandProperties;
 import com.netflix.hystrix.exception.HystrixRuntimeException;
@@ -13,13 +14,17 @@ import com.opower.rest.client.generator.hystrix.HystrixClient;
 import com.opower.rest.test.jetty.JettyRule;
 import com.opower.rest.test.resource.Frob;
 import com.opower.rest.test.resource.FrobResource;
+import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.util.Properties;
+import java.util.concurrent.TimeUnit;
 import org.junit.ClassRule;
 import org.junit.Test;
 
 import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.CoreMatchers.notNullValue;
 import static org.junit.Assert.assertThat;
+import static org.junit.Assert.fail;
 
 /**
  * Tests for the integration with the codahale-3-metrics-provider.
@@ -36,45 +41,47 @@ public class Codahale3MetricsIntegrationTesting {
     private static final String CLIENT_ID = "client.id";
     private static final String FROB_ID = "frobId";
     private static final int FROB_METHOD_COUNT = FrobResource.class.getDeclaredMethods().length;
+    private static final int TIMEOUT = (int)TimeUnit.MINUTES.toMillis(10L);
+    private AnnotatedFrobResource client;
+    private MetricRegistry metricRegistry;
 
     /**
      * Tests that metrics appear in the codahale MetricRegistry like they should.
      */
     @Test
     public void metricsAreRecorded() {
-        AnnotatedFrobResource client = new OpowerClient.Builder<>(AnnotatedFrobResource.class,
+        this.client = new OpowerClient.Builder<>(AnnotatedFrobResource.class,
                                                         new SimpleUriProvider("http://localhost:7000/"), 
                                                         CLIENT_ID)
                 .disableSensuPublishing()
                 .commandProperties(new ConfigurationCallback<HystrixCommandProperties.Setter>() {
                     @Override
                     public void configure(HystrixCommandProperties.Setter setter) {
-                        setter.withExecutionIsolationThreadTimeoutInMilliseconds(PORT);
+                        setter.withExecutionIsolationThreadTimeoutInMilliseconds(TIMEOUT);
                     }
                 })
                 .build();
         
-        client.findFrob(FROB_ID);
-        client.updateFrob(FROB_ID, FROB_ID);
-        client.createFrob(new Frob(FROB_ID));
-        client.frobString(FROB_ID);
+        this.client.findFrob(FROB_ID);
+        this.client.updateFrob(FROB_ID, FROB_ID);
+        this.client.createFrob(new Frob(FROB_ID));
+        this.client.frobString(FROB_ID);
 
         try {
-            client.frobJsonError();
+            this.client.frobJsonError();
         } catch (Exception ex) {
             // ignore
         }
 
         try {
-            client.frobErrorResponse();
+            this.client.frobErrorResponse();
         } catch (Exception e) {
             // ignore
         }
 
-        MetricRegistry metricRegistry = SharedMetricRegistries.getOrCreate(String.format("%s.client",SERVICE_NAME));
-        assertThat(metricRegistry.getTimers().size(), is(FROB_METHOD_COUNT));
-        assertThat(metricRegistry.getCounters().size(), is(1));
-
+        this.metricRegistry = SharedMetricRegistries.getOrCreate(String.format("%s.client",SERVICE_NAME));
+        assertThat(this.metricRegistry.getTimers().size(), is(FROB_METHOD_COUNT));
+        assertThat(this.metricRegistry.getCounters().size(), is(1));
 
         Properties props = new Properties();
         // force all the circuit breakers open
@@ -86,56 +93,44 @@ public class Codahale3MetricsIntegrationTesting {
         
         ConfigurationManager.loadProperties(props);
 
-        try {
-            client.findFrob(FROB_ID);
-            checkFrobMethod(metricRegistry, "findFrob");
-        } catch (HystrixRuntimeException e) {
-            assertThat(e.getFailureType(), is(HystrixRuntimeException.FailureType.SHORTCIRCUIT));
-        }
-        try {
-            client.updateFrob(FROB_ID, FROB_ID);
-            checkFrobMethod(metricRegistry, "updateFrob");
-        } catch (HystrixRuntimeException e) {
-            assertThat(e.getFailureType(), is(HystrixRuntimeException.FailureType.SHORTCIRCUIT));
-        }
-        try {
-            client.createFrob(new Frob(FROB_ID));
-            checkFrobMethod(metricRegistry, "createFrob");
-        } catch (HystrixRuntimeException e) {
-            assertThat(e.getFailureType(), is(HystrixRuntimeException.FailureType.SHORTCIRCUIT));
-        }
-        try {
-            client.frobString(FROB_ID);
-            checkFrobMethod(metricRegistry, "frobString");
-        } catch (HystrixRuntimeException e) {
-            assertThat(e.getFailureType(), is(HystrixRuntimeException.FailureType.SHORTCIRCUIT));
-        }
-        try {
-            client.frobJsonError();
-            checkFrobMethod(metricRegistry, "frobJsonError");
-        } catch (HystrixRuntimeException e) {
-            assertThat(e.getFailureType(), is(HystrixRuntimeException.FailureType.SHORTCIRCUIT));
-        }
-        try {
-            client.frobErrorResponse();
-            checkFrobMethod(metricRegistry, "frobErrorResponse");
-        } catch (HystrixRuntimeException e) {
-            assertThat(e.getFailureType(), is(HystrixRuntimeException.FailureType.SHORTCIRCUIT));
-        }
-
-
+        checkMetrics("findFrob", FROB_ID);
+        checkMetrics("updateFrob", FROB_ID, FROB_ID);
+        checkMetrics("createFrob", new Frob(FROB_ID));
+        checkMetrics("frobString", FROB_ID);
+        checkMetrics("frobJsonError");
+        checkMetrics("frobErrorResponse");
     }
 
-    private void checkFrobMethod(MetricRegistry metricRegistry, String methodName) {
+    /**
+     *
+     */
+    private void checkMetrics(String methodName, Object... args) {
+        try {
+            Class[] paramTypes = new Class[args.length];
+            for (int i = 0; i < args.length; i++) {
+                paramTypes[i] = args[i].getClass();
+            }
+            try {
+                FrobResource.class.getDeclaredMethod(methodName, paramTypes).invoke(this.client, args);
+            } catch (NoSuchMethodException | IllegalAccessException  ex) {
+                Throwables.propagate(ex);
+            } catch (InvocationTargetException ex) {
+                Throwables.propagate(ex.getTargetException());
+            }
+            fail();
+        } catch (HystrixRuntimeException e) {
+            assertThat(e.getFailureType(), is(HystrixRuntimeException.FailureType.SHORTCIRCUIT));
+            checkFrobMethod(methodName);
+        }
+    }
+
+    private void checkFrobMethod(String methodName) {
         String timerName = String.format("%s.%s", FrobResource.class.getCanonicalName(), methodName);
-        assertThat(metricRegistry.getTimers().get(timerName).getCount(), is(1L));
-        String exceptionCounterName = timerName + ".exceptions";
-        assertThat(metricRegistry.getCounters().get(exceptionCounterName).getCount(), is(1L));
+        assertThat(this.metricRegistry.getTimers().get(timerName), notNullValue());
+        String exceptionCounterName = timerName + "-Exception";
+        assertThat(this.metricRegistry.getCounters().get(exceptionCounterName), notNullValue());
         String gaugeName = timerName + ".circuitbreaker.open";
-        assertThat((int)metricRegistry.getGauges().get(gaugeName).getValue(), is(1));
-        String shortCircuitCounterName = timerName + ".shortcircuit.count";
-        assertThat(metricRegistry.getCounters().get(shortCircuitCounterName).getCount(), is(1L));
-                
+        assertThat((int)this.metricRegistry.getGauges().get(gaugeName).getValue(), is(1));
     }